Develop

+++ simple_interest.cpp
@@ -3,9 +3,9 @@ using namespace std;
 
 
 int main() {
-  float p = 1000;
-  float r = 4;
-float t = 2 ;
+  int p = 1000;
+  int r = 4;
+int t = 2 ;
 
 cout << (p * r * t) / 100.0  << endl;
 }
